Split component into URLRegister
