+++ src/component/molecules/URLRegister/URLRegister.css
@@ -1,0 +1,3 @@
+.com-m-URLRegister__container {
+  height: 64px;
+}
